api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
anthropic==0.25.7
httpx==0.24.1
scikit-learn==1.3.2
numpy==1.24.3
pydantic==2.5.0
python-dotenv==1.0.0
tqdm==4.66.1
